upd: chia ra nhieu components
