+++ components/nav-header/index.tsx
@@ -9,7 +9,7 @@ type Props = {};
 export default function Header({}: Props) {
   return (
     <header className="flex flex-row justify-between fixed top-0 left-0 right-0 px-5 py-5 text-white z-50">
-      <div className="flex flex-1 justify-start items-center max-w-8">
+      <div className="flex flex-1 justify-start items-center max-w-sm">
         <Logo />
         {/* <Image src="/images/logo.svg" width={50} height={50} alt="jaskdanj"/> */}
       </div>
