+++ icons/logo.tsx
@@ -8,7 +8,6 @@ export default function Logo({}: Props) {
       xmlns="http://www.w3.org/2000/svg"
       viewBox="0 0 1563.57 270.4"
       fill="white"
-      className="max-h-8"
     >
       <path
         d="M372.62,932q4.39,11.21-3.6,20l-69.2,68.4q-3.6,4.41,0,7.2l68.4,68.8q8.79,8.8,4.4,20a17.8,17.8,0,0,1-6.8,8,18.34,18.34,0,0,1-10.4,3.2h-15.2a16,16,0,0,1-12.8-5.6l-80-79.6a17.37,17.37,0,0,1-5.2-12.8v-11.2q0-6.79,5.2-12.8l80-79.6a16,16,0,0,1,12.8-5.6h15.6a17.5,17.5,0,0,1,10.2,3.2A18,18,0,0,1,372.62,932Z"
